Remove ReaderPostCellActions
