+++ WordPress/Classes/ViewRelated/Reader/ReaderPostCardCellViewModel.swift
@@ -233,5 +233,29 @@ struct ReaderPostCardCellViewModel {
         )
         WPAnalytics.trackReader(.postCardMoreTapped)
     }
+}
 
+enum ReaderActionsVisibility: Equatable {
+    case hidden
+    case visible(enabled: Bool)
+
+    static func == (lhs: ReaderActionsVisibility, rhs: ReaderActionsVisibility) -> Bool {
+        switch (lhs, rhs) {
+        case (.hidden, .hidden):
+            return true
+        case (.visible(let lenabled), .visible(let renabled)):
+            return lenabled == renabled
+        default:
+            return false
+        }
+    }
+
+    var isEnabled: Bool {
+        switch self {
+        case .hidden:
+            return false
+        case .visible(let enabled):
+            return enabled
+        }
+    }
 }
