--- WordPress/Classes/ViewRelated/Reader/ReaderSavedPostCellActions.swift
@@ -1,4 +1,0 @@
-// TODO: Delete when the reader improvements v1 (`readerImprovements`) flag is deleted
-protocol ReaderSavedPostCellActionsDelegate: AnyObject {
-    func willRemove(_ cell: OldReaderPostCardCell)
-}
